исправить ошибки при проверке в Postman

+++ controllers/cards.js
@@ -1,9 +1,10 @@
 const BadRequestError = require("../errors/bad-request-err");
 const NotFoundError = require("../errors/not-found-err");
+const ForbiddenError = require("../errors/forbidden-err");
 
 const Card = require("../models/card");
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       if (cards.length === 0) {
@@ -15,7 +16,7 @@ module.exports.getCards = (req, res) => {
     .catch(next);
 };
 
-module.exports.createNewCard = (req, res) => {
+module.exports.createNewCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => {
@@ -29,24 +30,23 @@ module.exports.createNewCard = (req, res) => {
     .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
         throw new NotFoundError("Нет карточки с данным id");
       }
-      res.send(card);
-    })
-    .catch((err) => {
-      if (err.name === "CastError") {
-        throw new BadRequestError("Невалидный id");
+      if (card.owner.toString() !== req.user._id.toString()) {
+        throw new ForbiddenError("Недостаточно прав");
       }
+      Card.findByIdAndRemove(req.params.cardId)
+        .then((deletedCard) => res.send(deletedCard))
+        .catch(next);
     })
     .catch(next);
 };
 
-module.exports.likeCard = (req, res) =>
+module.exports.likeCard = (req, res, next) =>
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -65,7 +65,7 @@ module.exports.likeCard = (req, res) =>
     })
     .catch(next);
 
-module.exports.dislikeCard = (req, res) =>
+module.exports.dislikeCard = (req, res, next) =>
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
